# admin, login, main 종합 완료

+++ DongHang_ComeFunny/src/main/java/com/DongHang_ComeFunny/www/controller/admin/AdminFreeBoardController.java
@@ -12,6 +12,7 @@ import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.FileSystemResource;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -19,6 +20,9 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.DongHang_ComeFunny.www.model.service.admin.AdminFreeBoardService;
+import com.DongHang_ComeFunny.www.model.vo.Admin;
+import com.DongHang_ComeFunny.www.model.vo.FreeComment;
+import com.DongHang_ComeFunny.www.model.vo.User;
 
 @Controller
 @RequestMapping("/admin/boards/freeBoard")
@@ -30,10 +34,18 @@ public class AdminFreeBoardController {
 	@RequestMapping("/list")
 	public ModelAndView viewFreeBoardList (	String searchKinds,
 											String searchText,
+											HttpSession session,
 											@RequestParam(required = false, defaultValue = "1")int cPage) {
 		
 		ModelAndView mav = new ModelAndView();
+		
+		Admin sessionAdmin= (Admin)session.getAttribute("adminLoginInfo");
+		User sessionUser =(User)session.getAttribute("logInInfo");
+		if(sessionAdmin != null) {
+			
 		int cntPerPage = 10;
+		
+		
 		
 		Map<String,Object> searchFreeBoard = new HashMap<>();
 		
@@ -47,13 +59,30 @@ public class AdminFreeBoardController {
 		mav.addObject("searchText", searchText);
 		mav.setViewName("admin/boards/freeBoard/list");
 		return mav;
+		}
+		 else if (sessionUser != null) {
+				mav.addObject("alertMsg", "관리자만 이용 가능합니다.");
+				mav.addObject("url", "/main");
+				mav.setViewName("common/result");
+				return mav;
+			} else {
+				mav.addObject("alertMsg", "로그인해 주세요.");
+				mav.addObject("url", "/admin/login");
+				mav.setViewName("common/result");
+				return mav;
+			}
 	}
 	
 		@RequestMapping("/delete")
-		public ModelAndView deleteFreeBoard(String[] fbNo) {
+		public ModelAndView deleteFreeBoard(HttpSession session,
+											String[] fbNo) {
 			
 			ModelAndView mav = new ModelAndView();
-					
+			
+			Admin sessionAdmin= (Admin)session.getAttribute("adminLoginInfo");
+			User sessionUser =(User)session.getAttribute("logInInfo");
+			if(sessionAdmin != null) {
+				
 					if(fbNo != null) {
 						adminFreeBoardService.deleteFreeBoard(fbNo);
 						mav.setViewName("redirect:/admin/boards/freeBoard/list");
@@ -62,13 +91,35 @@ public class AdminFreeBoardController {
 						mav.setViewName("redirect:/admin/boards/freeBoard/list");
 						return mav;
 					}
+			}else if (sessionUser != null) {
+				mav.addObject("alertMsg", "관리자만 이용 가능합니다.");
+				mav.addObject("url", "/main");
+				mav.setViewName("common/result");
+				return mav;
+			} else {
+				mav.addObject("alertMsg", "로그인해 주세요.");
+				mav.addObject("url", "/admin/login");
+				mav.setViewName("common/result");
+				return mav;
+			}
 		}
 		
+		// url : /board/freeview - GET
+		// parameter :
+		// int fbNo - 파라미터(name="fbNo") 값(value) 불러오기
 		@RequestMapping(value = "/view", method = RequestMethod.GET)
-		public ModelAndView freeView(int fbNo
+		public ModelAndView viewFreeBoard(HttpSession session,
+											int fbNo
 			) {
+			// 1. 모델앤뷰 객체 생성
 			ModelAndView mav = new ModelAndView();
-			Map<String, Object> commandMap = adminFreeBoardService.viewtFreeBoard(fbNo);
+			
+			Admin sessionAdmin= (Admin)session.getAttribute("adminLoginInfo");
+			User sessionUser =(User)session.getAttribute("logInInfo");
+			if(sessionAdmin != null) {
+				
+			// 2. 파라미터값 Service에 전달 후 Map에 저장
+			Map<String, Object> commandMap = adminFreeBoardService.viewFreeBoard(fbNo);
 			
 			//---------------- Service에서 처리한 데이터값 출력-------------------------
 			System.out.println("[controller] freeview - commandMap : " + commandMap);
@@ -89,6 +140,17 @@ public class AdminFreeBoardController {
 			}
 			// 6. 데이터 처리가 완료된 모델(VO)값과 보여질 페이지(jsp)를 반환한다.
 			return mav;
+			}else if (sessionUser != null) {
+				mav.addObject("alertMsg", "관리자만 이용 가능합니다.");
+				mav.addObject("url", "/main");
+				mav.setViewName("common/result");
+				return mav;
+			} else {
+				mav.addObject("alertMsg", "로그인해 주세요.");
+				mav.addObject("url", "/admin/login");
+				mav.setViewName("common/result");
+				return mav;
+			}
 		}
 		
 		
@@ -138,6 +200,27 @@ public class AdminFreeBoardController {
 			// 1. 파라미터값 Service에 전달 후 int 값 반환
 			//    ( 성공 : 1, 실패 : 0 )
 			return adminFreeBoardService.viewFreeCommentList(fbNo);
+		}
+		
+		@RequestMapping(value="/deleteReply", method = RequestMethod.PUT)
+		@ResponseBody
+		public int deleteReply(@RequestBody Map<String,Object> replyNo) {
+			
+			//-------------------RequestBody에서 요청한 값 제대로 받아왔는지 확인하기----------------------
+			System.out.println("[controller] updateReply map : "+replyNo);
+			System.out.println("[controller] updateReply fcno int: "+replyNo.get("replyNo"));
+			System.out.println("[controller] updateReply fccontent string : "+replyNo.get("replyText"));
+			//---------------------------------------------------------------------------------------------
+			
+			// 1. FreeComment fcNo에 Map에서 갖고온 Value(Object 타입)을 int형태로 바꾸기 
+			int fcNo = Integer.parseInt((String) replyNo.get("replyNo"));
+			// 2. FreeComment 빈 객체 생성
+			FreeComment freeComment = new FreeComment();
+			// 3. FreeComment에 fcNo, fcContent 값 지정
+			freeComment.setFcNo(fcNo);
+			// 4. 파라미터값 Service에 전달 후 int 값 반환
+			//    ( 성공 : 1, 실패 : 0 )
+			return adminFreeBoardService.deleteFreeComment(freeComment);
 		}
